Moved @tryghost/extract-api-key to @tryghost/mw-api-version-mismatch folder
fix https://linear.app/ghost/issue/ENG-2091/extract-api-key

- this is the only place we use this little snippet of code so we can
  just move it to the package to cut down on overhead for keeping a
  small piece of code
