--- ghost/extract-api-key/README.md
@@ -1,3 +1,0 @@
-# Extract Api Key
-
-Extracts API key value from requests in formats known to Ghost
