--- ghost/extract-api-key/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/extract-api-key');
