+++ ghost/mw-api-version-mismatch/package.json
@@ -24,6 +24,6 @@
     "sinon": "15.2.0"
   },
   "dependencies": {
-    "@tryghost/extract-api-key": "0.0.0"
+    "jsonwebtoken": "8.5.1"
   }
 }
